+++ talk/media/base/audiorenderer.h
@@ -41,7 +41,7 @@ class AudioRenderer {
     virtual void OnData(const void* audio_data,
                         int bits_per_sample,
                         int sample_rate,
-                        int number_of_channels,
+                        size_t number_of_channels,
                         size_t number_of_frames) = 0;
 
     // Called when the AudioRenderer is going away.
