+++ webrtc/audio/audio_sink.h
@@ -30,7 +30,7 @@ class AudioSinkInterface {
     Data(int16_t* data,
          size_t samples_per_channel,
          int sample_rate,
-         int channels,
+         size_t channels,
          uint32_t timestamp)
         : data(data),
           samples_per_channel(samples_per_channel),
@@ -41,7 +41,7 @@ class AudioSinkInterface {
     int16_t* data;               // The actual 16bit audio data.
     size_t samples_per_channel;  // Number of frames in the buffer.
     int sample_rate;             // Sample rate in Hz.
-    int channels;                // Number of channels in the audio data.
+    size_t channels;             // Number of channels in the audio data.
     uint32_t timestamp;          // The RTP timestamp of the first sample.
   };
